pythonProject/requirements.txt
DateTime==4.3
kafka-python==2.0.2
pytz==2021.1
zope.interface==5.4.0
